Fixed Shell_getCurrentTime() to not overflow at 768 seconds on pre-iPhone-4 systems iPhone SDK 4.0.1 fixes
git-svn-id: http://ludobloom.com/svn/StemLibProjects@635 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ eaglshell/trunk/source/eaglshell/EAGLShell.m
@@ -73,7 +73,7 @@ double Shell_getCurrentTime() {
 	if (timebaseInfo.denom == 0) {
 		mach_timebase_info(&timebaseInfo);
 	}
-	return mach_absolute_time() * timebaseInfo.numer / timebaseInfo.denom * 0.000000001;
+	return mach_absolute_time() * (double) timebaseInfo.numer / timebaseInfo.denom * 0.000000001;
 }
 
 const char * Shell_getResourcePath() {
